requirements.txt
asgiref==3.8.1
Django==3.2.16
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.15.1
pillow==10.3.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.11.0
